+++ src/components/echo-finale/index.js
@@ -1,0 +1,7 @@
+import "./echo-finale.css";
+
+export {
+  ECHO_FINALE_LOOP,
+  echoFinaleState,
+  echoFinaleMethods,
+} from "./echo-finale.js";